% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{access_api}
\alias{access_api}
\title{Query the thenmap API}
\usage{
access_api(api, dataset_name, module, date, req_parameters = NULL)
}
\arguments{
\item{api}{A thenmapAPI object}

\item{dataset_name}{The name of the dataset for which to retrieve the data}

\item{module}{The API module to be queried, one of 'data', 'info', 'geo', 'svg'}

\item{date}{A character vector with the for which to retrieve the data, must be one of the formats 'YYYY', 'YYYY-MM', 'YYYY-MM-DD'}

\item{req_parameters}{List of request parameters}
}
\value{
The httr2 response object
}
\description{
Query the thenmap API
}
\examples{
access_api(api, "se-7", "data", "2015")
}
