% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{get_data}
\alias{get_data}
\title{Get data from the thenmap API}
\usage{
get_data(
  api,
  dataset_name,
  date,
  language = NULL,
  data_format = "json",
  return_object = TRUE
)
}
\arguments{
\item{api}{A thenmapAPI object}

\item{dataset_name}{The name of the dataset for which to retrieve the data}

\item{date}{A character vector with the for which to retrieve the data, must be one of the formats 'YYYY', 'YYYY-MM', 'YYYY-MM-DD'}

\item{language}{Optional: The two-letter language code of the language in which to retrieve the data}

\item{data_format}{The format in which to request the data from the API, one of 'json', 'csv'. Default: json}

\item{return_object}{Whether to return on object suitable for further processing, in this case a data.frame. If FALSE, the response object returned by httr2 is returned. Default: TRUE}
}
\value{
A data.frame if return_object==TRUE, a response object if return_object==FALSE
}
\description{
Get data from the thenmap API
}
\examples{
api <- thenmapsAPI(version="v2")
get_data(api=api, dataset_name="se-7", date="2015")

}
\references{
\url{http://api.thenmap.net/doc/v2/#data}
}
