% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{get_geo}
\alias{get_geo}
\title{Get geo data from the thenmap API}
\usage{
get_geo(
  api,
  dataset_name,
  date,
  language = NULL,
  geo_type = "geojson",
  geo_crs = "wgs84",
  geo_props = NULL,
  return_object = TRUE
)
}
\arguments{
\item{api}{A thenmapAPI object}

\item{dataset_name}{The name of the dataset for which to retrieve the data}

\item{date}{A character vector with the for which to retrieve the data, must be one of the formats 'YYYY', 'YYYY-MM', 'YYYY-MM-DD'}

\item{language}{Optional: The two-letter language code of the language in which to retrieve the data}

\item{geo_type}{The format in which to request the geodata from the API, one of 'geojson', 'topojson'. Default: geojson}

\item{geo_crs}{The coordinate reference system to use for the geodata. Default: wgs84}

\item{geo_props}{A character vector containing all comma-separated fields from the data module to be included in the shape properties}

\item{return_object}{Whether to return on object suitable for further processing, in this case an sf object. If FALSE, the response object returned by httr2 is returned. Default: TRUE}
}
\value{
An sf object if return_object==TRUE, a response object if return_object==FALSE
}
\description{
Get geo data from the thenmap API
}
\examples{
api <- thenmapsAPI(version="v2")
get_geo(api=api, dataset_name="se-7", date="2015")

}
\references{
\url{http://api.thenmap.net/doc/v2/#geo}
}
