% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{get_info}
\alias{get_info}
\title{Get metadata for a dataset from the thenmap API}
\usage{
get_info(api, dataset_name, date, language = NULL, return_object = TRUE)
}
\arguments{
\item{api}{A thenmapAPI object}

\item{dataset_name}{The name of the dataset for which to retrieve the data}

\item{date}{A character vector with the for which to retrieve the data, must be one of the formats 'YYYY', 'YYYY-MM', 'YYYY-MM-DD'}

\item{language}{Optional: The two-letter language code of the language in which to retrieve the data}

\item{return_object}{Whether to return on object suitable for further processing, in this case a list. If FALSE, the response object returned by httr2 is returned. Default: TRUE}
}
\value{
A list if return_object==TRUE, a response object if return_object==FALSE
}
\description{
Get metadata for a dataset from the thenmap API
}
\examples{
api <- thenmapsAPI(version="v2")
get_info(api=api, dataset_name="se-7", date="2015")

}
\references{
\url{http://api.thenmap.net/doc/v2/#info}
}
