% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{get_svg}
\alias{get_svg}
\title{Get svg representations from the thenmap API}
\usage{
get_svg(
  api,
  dataset_name,
  date,
  language = NULL,
  svg_proj = NULL,
  svg_width = NULL,
  svg_height = NULL,
  svg_props = NULL,
  return_object = TRUE
)
}
\arguments{
\item{api}{A thenmapAPI object}

\item{dataset_name}{The name of the dataset for which to retrieve the data}

\item{date}{A character vector with the for which to retrieve the data, must be one of the formats 'YYYY', 'YYYY-MM', 'YYYY-MM-DD'}

\item{language}{Optional: The two-letter language code of the language in which to retrieve the data}

\item{svg_proj}{The projection to use for the dataset when represented in the SVG. Defaults to the first recommendation if not provided}

\item{svg_width}{The with of the SVG in pixels}

\item{svg_height}{The height of the SVG in pixels}

\item{svg_props}{A character vector containing all comma-separated fields from the data module to be included as data attributes}

\item{return_object}{Whether to return on object suitable for further processing, in this case an svg_obj. If FALSE, the response object returned by httr2 is returned. Default: TRUE}
}
\value{
An svg_obj if return_object==TRUE, a response object if return_object==FALSE
}
\description{
Get svg representations from the thenmap API
}
\examples{
api <- thenmapsAPI(version="v2")
get_svg(api=api, dataset_name="se-7", date="2015")

}
\references{
\url{http://api.thenmap.net/doc/v2/#svg}
}
