% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thenmapsAPI.R
\name{thenmapsAPI}
\alias{thenmapsAPI}
\title{Create thenmapsAPI instance}
\usage{
thenmapsAPI(version = "v2")
}
\arguments{
\item{version}{The version of the API to use, either "v1" or "v2"}
}
\value{
A list of class thenmapAPI with the attributes base_url, version, datasets, modules, valid_projections, data_formats, and geo_type
}
\description{
Create thenmapsAPI instance
}
\examples{
api <- thenmapsAPI(version="v2")

}
\references{
\url{https://www.thenmap.net/}
}
